requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework-simplejwt==5.5.0
django-autoslug==1.9.9
django-cities==0.6.2
django-cities-light==3.10.1
django-countries==7.6.1
django-model-utils==5.0.0
django-modeltranslation==0.19.12
django-social-share==2.3.0
django-tinymce==4.1.0
django-view-breadcrumbs==2.5.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
pillow==11.1.0
progressbar2==4.5.0
python-utils==3.9.1
pytz==2025.1
setuptools==75.8.1
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tqdm==4.67.1
transliterate==1.10.2
typing_extensions==4.12.2
Unidecode==1.3.8
django_recaptcha